Create an ellmer chat with the data of the omop trained model

Create an ellmer chat with the data of the omop trained model

Usage

ellmerChat(name = "omop_assistant", chat = ellmer::chat_google_gemini(), ...)

Arguments

name

Store name.

chat

A chat object.

...

Arguments passed to ragnar::ragnar_register_tool_retrieve()

Value

The chat object with the trained store and prompt.

Examples

# \donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)
#> Error in storeCreate(embed = function(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),     name = "my_omop_assistant"): could not find function "storeCreate"

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)
#> Error in ellmerChat(name = "my_omop_assistant", chat = ellmer::chat_google_gemini(),     top_k = 10L): ✖ my_omop_assistant does not exist, please use `downloadModel()` or
#>   `trainModel()` to create the model first.

chat$chat("How to create an acetaminophen cohort?")
#> Error: object 'chat' not found
# }