Create a store

Create a store

Usage

storeCreate(embed, name = "omop_assistant", overwrite = FALSE)

Arguments

embed

A function that is called with a character vector and returns a matrix of embeddings. Note this function will be serialised and then deserialised in new R sessions, so it cannot reference to any objects in the global or parent environments. Make sure to namespace all function calls with ::. If additional R objects must be available in the function, you can optionally supply a carrier::crate() with packaged data. It can also be NULL for stores that don't need to embed their texts, for example, if only using FTS algorithms such as ragnar_retrieve_bm25().

name

Store name.

overwrite

Whether to overwrite a preexisting store.

Value

String with the path to the store.

Examples

# \donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)
#> ℹ `OMOP_DATA_FOLDER` environment variable is not set, using temp directory.
#> ℹ Using already created storage in:
#>   /tmp/RtmpY9Gmf0/OMOP_DATA_FOLDER/my_omop_assistant.duckdb.
#> [1] "/tmp/RtmpY9Gmf0/OMOP_DATA_FOLDER/my_omop_assistant.duckdb"

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)
#> ℹ Set system prompt.
#> Using model = "gemini-2.5-flash".
#> Error in default_google_credentials(api_key, gemini = TRUE): The package "gargle" is required for Google authentication

chat$chat("How to create an acetaminophen cohort?")
#> Error: object 'chat' not found
# }