Create an ellmer chat with the data of the omop trained model

Create an ellmer chat with the data of the omop trained model

Usage

ellmerChat(name = "omop_assistant", chat = ellmer::chat_google_gemini(), ...)

Arguments

name

Store name.

chat

A chat object.

...

Arguments passed to ragnar::ragnar_register_tool_retrieve()

Value

The chat object with the trained store and prompt.

Examples

# \donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)
#> ℹ `OMOP_DATA_FOLDER` environment variable is not set, using temp directory.
#> ℹ Retrieving information links.
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/adding_features.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/cdm_reference.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/creating_cohorts.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/dbplyr_packages.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/exploring_the_cdm.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/intro.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/omop.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/tidyverse_expressions.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/tidyverse_verbs.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/working_with_cohorts.html>
#> ℹ Reading information from
#>   <https://oxford-pharmacoepi.github.io/Tidy-R-programming-with-OMOP/working_with_databases_from_r.html>
#> ℹ Reading information from <https://darwin-eu.github.io/CDMConnector/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/articles/a01_getting-started.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/articles/a02_cohorts.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/articles/a03_dbplyr.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/articles/a04_DBI_connection_examples.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/articles/a06_using_cdm_attributes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/pare_report.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/reference/cdmFromCon.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/reference/eunomiaDir.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CDMConnector/reference/pipe.html>
#> ℹ Reading information from <https://darwin-eu.github.io/omopgenerics/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/cdm_reference.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/codelists.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/cohorts.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/expanding_omopgenerics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/logging.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/reexport.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/articles/summarised_result.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/achillesColumns.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/achillesTables.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/attrition.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/cdmFromTables.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/cohortColumns.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/cohortCount.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/cohortTables.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/newCodelist.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/newCohortTable.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/newConceptSetExpression.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/omopColumns.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/omopTables.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/omopgenerics/reference/settings.html>
#> ℹ Reading information from <https://ohdsi.github.io/CohortConstructor/>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a00_introduction.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a01_building_base_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a02_cohort_table_requirements.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a03_require_demographics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a04_require_intersections.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a05_update_cohort_start_end.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a06_concatanate_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a07_filter_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a08_split_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a09_combine_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a10_match_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a11_benchmark.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/articles/a12_behind_the_scenes.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/authors.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/news/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/conceptCohort.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/padCohortEnd.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/requireCohortIntersect.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/requireDemographics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/requireInDateRange.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/CohortConstructor/reference/unionCohorts.html>
#> ℹ Reading information from <https://darwin-eu.github.io/visOmopResults/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/articles/a01_tables.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/articles/a02_plots.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/barPlot.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/mockSummarisedResult.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/themeVisOmop.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/visOmopTable.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/visOmopResults/reference/visTable.html>
#> ℹ Reading information from <https://ohdsi.github.io/PhenotypeR/>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/CodelistDiagnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/CohortDiagnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/DatabaseDiagnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/PhenotypeDiagnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/PhenotypeExpectations.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/PopulationDisgnostics.html>
#> ✖ Failed to read markdown in
#>   <https://ohdsi.github.io/PhenotypeR/articles/PopulationDisgnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/articles/ShinyDiagnostics.html>
#> ℹ Reading information from <https://ohdsi.github.io/PhenotypeR/authors.html>
#> ℹ Reading information from <https://ohdsi.github.io/PhenotypeR/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/reference/getCohortExpectations.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/reference/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/reference/phenotypeDiagnostics.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/PhenotypeR/reference/shinyDiagnostics.html>
#> ℹ Reading information from <https://ohdsi.github.io/OmopViewer/>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/background.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/dynamic_app.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/edit_static_content.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/export_static_app.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/extending_omopviewer.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/articles/theming.html>
#> ℹ Reading information from <https://ohdsi.github.io/OmopViewer/authors.html>
#> ℹ Reading information from <https://ohdsi.github.io/OmopViewer/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/reference/exportStaticApp.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/reference/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/OmopViewer/reference/launchDynamicApp.html>
#> ℹ Reading information from <https://darwin-eu.github.io/DrugUtilisation/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/create_cohorts.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/daily_dose_calculation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/drug_restart.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/drug_utilisation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/indication.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/mock_data.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/summarise_treatments.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/articles/treatment_discontinuation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/generateIngredientCohortSet.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/mockDrugUtilisation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/requireIsFirstDrugEntry.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/requireObservationBeforeDrug.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/summariseDrugUtilisation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugUtilisation/reference/summariseIndication.html>
#> ℹ Reading information from <https://darwin-eu.github.io/IncidencePrevalence/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a01_Introduction_to_IncidencePrevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a02_Creating_denominator_populations.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a03_Creating_target_denominator_populations.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a04_Calculating_prevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a05_Calculating_incidence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a06_Working_with_IncidencePrevalence_Results.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/articles/a07_benchmark.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/estimateIncidence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/estimatePeriodPrevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/estimatePointPrevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/generateDenominatorCohortSet.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/mockIncidencePrevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/plotIncidence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/IncidencePrevalence/reference/plotPrevalence.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DaysSupply.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DiagnosticsSummary.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DrugDose.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DrugDuration.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DrugSig.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DrugSourceConcepts.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/DrugTypes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/IntroductionToDrugExposureDiagnostics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/Missingness.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/Quantity.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/Routes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/articles/VerbatimEndDate.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/reference/executeChecks.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/reference/mockDrugExposure.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/reference/pipe.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/DrugExposureDiagnostics/reference/writeResultToDisk.html>
#> ℹ Reading information from <https://ohdsi.github.io/MeasurementDiagnostics/>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/articles/summariseMeasurementUse.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/authors.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/reference/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/reference/plotMeasurementTimings.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/MeasurementDiagnostics/reference/summariseMeasurementUse.html>
#> ℹ Reading information from <https://darwin-eu.github.io/PatientProfiles/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/articles/cohort-intersect.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/articles/concept-intersect.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/articles/demographics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/articles/summarise.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/articles/table-intersect.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addAge.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addCohortIntersectCount.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addCohortIntersectDate.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addCohortIntersectDays.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addCohortIntersectFlag.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addPriorObservation.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/addSex.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/mockDisconnect.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/PatientProfiles/reference/mockPatientProfiles.html>
#> ℹ Reading information from <https://darwin-eu.github.io/CohortCharacteristics/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/articles/summarise_characteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/articles/summarise_cohort_entries.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/articles/summarise_cohort_overlap.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/articles/summarise_cohort_timing.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/articles/summarise_large_scale_characteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/mockCohortCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotCohortAttrition.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotCohortCount.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotCohortOverlap.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotCohortTiming.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotComparedLargeScaleCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/plotLargeScaleCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseCohortAttrition.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseCohortCount.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseCohortOverlap.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseCohortTiming.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/summariseLargeScaleCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableCharacteristics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableCohortAttrition.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableCohortCount.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableCohortOverlap.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableCohortTiming.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CohortCharacteristics/reference/tableTopLargeScaleCharacteristics.html>
#> ℹ Reading information from <https://OHDSI.github.io/OmopSketch/>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/characterisation.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/database_characteristics.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/missing_data.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/summarise_clinical_tables_records.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/summarise_concept_id_counts.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/articles/summarise_observation_period.html>
#> ℹ Reading information from <https://OHDSI.github.io/OmopSketch/authors.html>
#> ℹ Reading information from <https://OHDSI.github.io/OmopSketch/index.html>
#> ℹ Reading information from <https://OHDSI.github.io/OmopSketch/news/index.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/index.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/plotConceptSetCounts.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/plotInObservation.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/plotObservationPeriod.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/plotRecordCount.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseClinicalRecords.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseConceptSetCounts.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseInObservation.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseObservationPeriod.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseOmopSnapshot.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/summariseRecordCount.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/tableClinicalRecords.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/tableObservationPeriod.html>
#> ℹ Reading information from
#>   <https://OHDSI.github.io/OmopSketch/reference/tableOmopSnapshot.html>
#> ℹ Reading information from <https://darwin-eu.github.io/CodelistGenerator/>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a01_GettingOmopCdmVocabularies.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a02_ExploreCDMvocabulary.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a03_GenerateCandidateCodelist.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a04_GenerateVocabularyBasedCodelist.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a04b_icd_codes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a05_ExtractCodelistFromJSONfile.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a06_CreateSubsetsFromCodelist.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/articles/a07_RunCodelistDiagnostics.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/getCandidateCodes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/getDrugIngredientCodes.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/getVocabVersion.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/summariseCodeUse.html>
#> ℹ Reading information from
#>   <https://darwin-eu.github.io/CodelistGenerator/reference/tableCodeUse.html>
#> ℹ Reading information from <https://darwin-eu-dev.github.io/CohortSurvival/>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/articles/a01_Single_event_of_interest.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/articles/a02_Competing_risk_survival.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/articles/a03_Further_survival_analyses.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/authors.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/news/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/estimateCompetingRiskSurvival.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/estimateSingleEventSurvival.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/index.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/mockMGUS2cdm.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/pipe.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/plotSurvival.html>
#> ℹ Reading information from
#>   <https://darwin-eu-dev.github.io/CohortSurvival/reference/tableSurvival.html>
#> ℹ Reading information from <https://ohdsi.github.io/omock/>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/articles/a01_Creating_synthetic_clinical_tables.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/articles/a02_Creating_synthetic_cohorts.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/articles/a03_Creating_a_synthetic_vocabulary.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/articles/a04_Building_a_bespoke_mock_cdm.html>
#> ℹ Reading information from <https://ohdsi.github.io/omock/authors.html>
#> ℹ Reading information from <https://ohdsi.github.io/omock/index.html>
#> ℹ Reading information from <https://ohdsi.github.io/omock/news/index.html>
#> ℹ Reading information from <https://ohdsi.github.io/omock/reference/index.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/reference/mockObservationPeriod.html>
#> ℹ Reading information from
#>   <https://ohdsi.github.io/omock/reference/mockPerson.html>
#> ✔ Information retrieved from 294 links.
#> ℹ Embeding retrieved information.
#> ⠙ 1/294 ETA:  2h | Embeding 
#> ⠹ 2/294 ETA:  6h | Embeding 
#> ⠸ 3/294 ETA:  8h | Embeding 
#> ⠼ 4/294 ETA:  7h | Embeding 
#> ⠴ 5/294 ETA:  9h | Embeding 
#> ⠦ 6/294 ETA:  8h | Embeding 
#> ⠧ 7/294 ETA:  7h | Embeding 
#> ⠇ 8/294 ETA:  6h | Embeding 
#> ⠏ 9/294 ETA:  9h | Embeding 
#> ⠋ 10/294 ETA:  9h | Embeding 
#> ⠙ 11/294 ETA:  8h | Embeding 
#> ⠹ 12/294 ETA:  8h | Embeding 
#> ⠸ 13/294 ETA:  8h | Embeding 
#> ⠼ 14/294 ETA:  8h | Embeding 
#> ⠴ 15/294 ETA:  8h | Embeding 
#> ⠦ 16/294 ETA:  8h | Embeding 
#> ⠧ 17/294 ETA:  8h | Embeding 
#> ⠇ 18/294 ETA:  7h | Embeding 
#> ⠏ 19/294 ETA:  7h | Embeding 
#> ⠋ 20/294 ETA:  7h | Embeding 
#> ⠙ 21/294 ETA:  7h | Embeding 
#> ⠹ 22/294 ETA:  6h | Embeding 
#> ⠸ 23/294 ETA:  6h | Embeding 
#> ⠼ 24/294 ETA:  6h | Embeding 
#> ⠴ 25/294 ETA:  6h | Embeding 
#> ⠦ 27/294 ETA:  5h | Embeding 
#> ⠧ 28/294 ETA:  5h | Embeding 
#> ⠇ 29/294 ETA:  5h | Embeding 
#> ⠏ 30/294 ETA:  5h | Embeding 
#> ⠋ 31/294 ETA:  5h | Embeding 
#> ⠙ 32/294 ETA:  5h | Embeding 
#> ⠹ 33/294 ETA:  5h | Embeding 
#> ⠸ 34/294 ETA:  5h | Embeding 
#> ⠼ 35/294 ETA:  5h | Embeding 
#> ⠴ 36/294 ETA:  5h | Embeding 
#> ⠦ 37/294 ETA:  5h | Embeding 
#> ⠧ 38/294 ETA:  5h | Embeding 
#> ⠇ 39/294 ETA:  5h | Embeding 
#> ⠏ 41/294 ETA:  4h | Embeding 
#> ⠋ 42/294 ETA:  4h | Embeding 
#> ⠙ 43/294 ETA:  4h | Embeding 
#> ⠹ 45/294 ETA:  4h | Embeding 
#> ⠸ 47/294 ETA:  4h | Embeding 
#> ⠼ 49/294 ETA:  4h | Embeding 
#> ⠴ 50/294 ETA:  4h | Embeding 
#> ⠦ 51/294 ETA:  4h | Embeding 
#> ⠧ 52/294 ETA:  4h | Embeding 
#> ⠇ 53/294 ETA:  3h | Embeding 
#> ⠏ 54/294 ETA:  4h | Embeding 
#> ⠋ 55/294 ETA:  4h | Embeding 
#> ⠙ 56/294 ETA:  3h | Embeding 
#> ⠹ 57/294 ETA:  3h | Embeding 
#> ⠸ 58/294 ETA:  3h | Embeding 
#> ⠼ 59/294 ETA:  3h | Embeding 
#> ⠴ 60/294 ETA:  3h | Embeding 
#> ⠦ 61/294 ETA:  3h | Embeding 
#> ⠧ 62/294 ETA:  3h | Embeding 
#> ⠇ 63/294 ETA:  3h | Embeding 
#> ⠏ 64/294 ETA:  3h | Embeding 
#> ⠋ 65/294 ETA:  3h | Embeding 
#> ⠙ 66/294 ETA:  3h | Embeding 
#> ⠹ 67/294 ETA:  3h | Embeding 
#> ⠸ 68/294 ETA:  3h | Embeding 
#> ⠼ 69/294 ETA:  3h | Embeding 
#> ⠴ 70/294 ETA:  3h | Embeding 
#> ⠦ 71/294 ETA:  3h | Embeding 
#> ⠧ 72/294 ETA:  3h | Embeding 
#> ⠇ 73/294 ETA:  3h | Embeding 
#> ⠏ 74/294 ETA:  3h | Embeding 
#> ⠋ 75/294 ETA:  3h | Embeding 
#> ⠙ 76/294 ETA:  3h | Embeding 
#> ⠹ 77/294 ETA:  3h | Embeding 
#> ⠸ 78/294 ETA:  3h | Embeding 
#> ⠼ 79/294 ETA:  3h | Embeding 
#> ⠴ 80/294 ETA:  3h | Embeding 
#> ⠦ 81/294 ETA:  3h | Embeding 
#> ⠧ 82/294 ETA:  3h | Embeding 
#> ⠇ 83/294 ETA:  3h | Embeding 
#> ⠏ 84/294 ETA:  3h | Embeding 
#> ⠋ 85/294 ETA:  3h | Embeding 
#> ⠙ 86/294 ETA:  3h | Embeding 
#> ⠹ 87/294 ETA:  3h | Embeding 
#> ⠸ 88/294 ETA:  3h | Embeding 
#> ⠼ 89/294 ETA:  3h | Embeding 
#> ⠴ 90/294 ETA:  3h | Embeding 
#> ⠦ 91/294 ETA:  3h | Embeding 
#> ⠧ 92/294 ETA:  3h | Embeding 
#> ⠇ 93/294 ETA:  3h | Embeding 
#> ⠏ 94/294 ETA:  3h | Embeding 
#> ⠋ 95/294 ETA:  3h | Embeding 
#> ⠙ 96/294 ETA:  3h | Embeding 
#> ⠹ 98/294 ETA:  3h | Embeding 
#> ⠸ 99/294 ETA:  3h | Embeding 
#> ⠼ 100/294 ETA:  3h | Embeding 
#> ⠴ 101/294 ETA:  3h | Embeding 
#> ⠦ 102/294 ETA:  2h | Embeding 
#> ⠧ 103/294 ETA:  2h | Embeding 
#> ⠇ 104/294 ETA:  2h | Embeding 
#> ⠏ 105/294 ETA:  2h | Embeding 
#> ⠋ 106/294 ETA:  2h | Embeding 
#> ⠙ 107/294 ETA:  2h | Embeding 
#> ⠹ 108/294 ETA:  2h | Embeding 
#> ⠸ 109/294 ETA:  2h | Embeding 
#> ⠼ 110/294 ETA:  2h | Embeding 
#> ⠴ 111/294 ETA:  2h | Embeding 
#> ⠦ 112/294 ETA:  2h | Embeding 
#> ⠧ 113/294 ETA:  2h | Embeding 
#> ⠇ 114/294 ETA:  2h | Embeding 
#> ⠏ 115/294 ETA:  2h | Embeding 
#> ⠋ 116/294 ETA:  2h | Embeding 
#> ⠙ 117/294 ETA:  3h | Embeding 
#> ⠹ 118/294 ETA:  3h | Embeding 
#> ⠸ 119/294 ETA:  3h | Embeding 
#> ⠼ 120/294 ETA:  3h | Embeding 
#> ⠴ 121/294 ETA:  2h | Embeding 
#> ⠦ 122/294 ETA:  2h | Embeding 
#> ⠧ 123/294 ETA:  2h | Embeding 
#> ⠇ 124/294 ETA:  2h | Embeding 
#> ⠏ 125/294 ETA:  2h | Embeding 
#> ⠋ 126/294 ETA:  2h | Embeding 
#> ⠙ 127/294 ETA:  2h | Embeding 
#> ⠹ 128/294 ETA:  2h | Embeding 
#> ⠸ 129/294 ETA:  2h | Embeding 
#> ⠼ 130/294 ETA:  2h | Embeding 
#> ⠴ 131/294 ETA:  2h | Embeding 
#> ⠦ 132/294 ETA:  2h | Embeding 
#> ⠧ 133/294 ETA:  2h | Embeding 
#> ⠇ 134/294 ETA:  2h | Embeding 
#> ⠏ 135/294 ETA:  2h | Embeding 
#> ⠋ 136/294 ETA:  2h | Embeding 
#> ⠙ 137/294 ETA:  2h | Embeding 
#> ⠹ 138/294 ETA:  2h | Embeding 
#> ⠸ 139/294 ETA:  2h | Embeding 
#> ⠼ 140/294 ETA:  2h | Embeding 
#> ⠴ 141/294 ETA:  2h | Embeding 
#> ⠦ 142/294 ETA:  2h | Embeding 
#> ⠧ 143/294 ETA:  2h | Embeding 
#> ⠇ 144/294 ETA:  2h | Embeding 
#> ⠏ 145/294 ETA:  2h | Embeding 
#> ⠋ 146/294 ETA:  2h | Embeding 
#> ⠙ 147/294 ETA:  2h | Embeding 
#> ⠹ 148/294 ETA:  2h | Embeding 
#> ⠸ 149/294 ETA:  2h | Embeding 
#> ⠼ 150/294 ETA:  2h | Embeding 
#> ⠴ 151/294 ETA:  2h | Embeding 
#> ⠦ 152/294 ETA:  2h | Embeding 
#> ⠧ 153/294 ETA:  2h | Embeding 
#> ⠇ 154/294 ETA:  2h | Embeding 
#> ⠏ 155/294 ETA:  2h | Embeding 
#> ⠋ 156/294 ETA:  2h | Embeding 
#> ⠙ 157/294 ETA:  2h | Embeding 
#> ⠹ 158/294 ETA:  2h | Embeding 
#> ⠸ 159/294 ETA:  2h | Embeding 
#> ⠼ 160/294 ETA:  2h | Embeding 
#> ⠴ 161/294 ETA:  2h | Embeding 
#> ⠦ 162/294 ETA:  2h | Embeding 
#> ⠧ 163/294 ETA:  2h | Embeding 
#> ⠇ 164/294 ETA:  2h | Embeding 
#> ⠏ 165/294 ETA:  2h | Embeding 
#> ⠋ 166/294 ETA:  2h | Embeding 
#> ⠙ 167/294 ETA:  2h | Embeding 
#> ⠹ 168/294 ETA:  2h | Embeding 
#> ⠸ 169/294 ETA:  2h | Embeding 
#> ⠼ 171/294 ETA:  2h | Embeding 
#> ⠴ 172/294 ETA:  2h | Embeding 
#> ⠦ 173/294 ETA:  2h | Embeding 
#> ⠧ 174/294 ETA:  2h | Embeding 
#> ⠇ 175/294 ETA:  2h | Embeding 
#> ⠏ 176/294 ETA:  1h | Embeding 
#> ⠋ 177/294 ETA:  1h | Embeding 
#> ⠙ 178/294 ETA:  1h | Embeding 
#> ⠹ 179/294 ETA:  1h | Embeding 
#> ⠸ 180/294 ETA:  1h | Embeding 
#> ⠼ 181/294 ETA:  1h | Embeding 
#> ⠴ 182/294 ETA:  1h | Embeding 
#> ⠦ 183/294 ETA:  1h | Embeding 
#> ⠧ 184/294 ETA:  1h | Embeding 
#> ⠇ 185/294 ETA:  1h | Embeding 
#> ⠏ 186/294 ETA:  1h | Embeding 
#> ⠋ 187/294 ETA:  1h | Embeding 
#> ⠙ 188/294 ETA:  1h | Embeding 
#> ⠹ 189/294 ETA:  1h | Embeding 
#> ⠸ 190/294 ETA:  1h | Embeding 
#> ⠼ 191/294 ETA:  1h | Embeding 
#> ⠴ 192/294 ETA:  1h | Embeding 
#> ⠦ 193/294 ETA:  1h | Embeding 
#> ⠧ 194/294 ETA:  1h | Embeding 
#> ⠇ 195/294 ETA:  1h | Embeding 
#> ⠏ 196/294 ETA:  1h | Embeding 
#> ⠋ 197/294 ETA:  1h | Embeding 
#> ⠙ 198/294 ETA:  1h | Embeding 
#> ⠹ 199/294 ETA:  1h | Embeding 
#> ⠸ 200/294 ETA:  1h | Embeding 
#> ⠼ 201/294 ETA:  1h | Embeding 
#> ⠴ 202/294 ETA:  1h | Embeding 
#> ⠦ 203/294 ETA:  1h | Embeding 
#> ⠧ 204/294 ETA:  1h | Embeding 
#> ⠇ 205/294 ETA:  1h | Embeding 
#> ⠏ 206/294 ETA:  1h | Embeding 
#> ⠋ 207/294 ETA:  1h | Embeding 
#> ⠙ 208/294 ETA:  1h | Embeding 
#> ⠹ 209/294 ETA:  1h | Embeding 
#> ⠸ 210/294 ETA:  1h | Embeding 
#> ⠼ 211/294 ETA:  1h | Embeding 
#> ⠴ 212/294 ETA:  1h | Embeding 
#> ⠦ 213/294 ETA:  1h | Embeding 
#> ⠧ 214/294 ETA:  1h | Embeding 
#> ⠇ 215/294 ETA:  1h | Embeding 
#> ⠏ 216/294 ETA:  1h | Embeding 
#> ⠋ 217/294 ETA:  1h | Embeding 
#> ⠙ 218/294 ETA:  1h | Embeding 
#> ⠹ 219/294 ETA:  1h | Embeding 
#> ⠸ 220/294 ETA:  1h | Embeding 
#> ⠼ 221/294 ETA:  1h | Embeding 
#> ⠴ 222/294 ETA:  1h | Embeding 
#> ⠦ 223/294 ETA:  1h | Embeding 
#> ⠧ 224/294 ETA:  1h | Embeding 
#> ⠇ 225/294 ETA: 49m | Embeding 
#> ⠏ 226/294 ETA: 48m | Embeding 
#> ⠋ 227/294 ETA: 48m | Embeding 
#> ⠙ 228/294 ETA: 47m | Embeding 
#> ⠹ 229/294 ETA: 47m | Embeding 
#> ⠸ 230/294 ETA: 47m | Embeding 
#> ⠼ 231/294 ETA:  1h | Embeding 
#> ⠴ 232/294 ETA:  1h | Embeding 
#> ⠦ 233/294 ETA:  1h | Embeding 
#> ⠧ 234/294 ETA:  1h | Embeding 
#> ⠇ 235/294 ETA:  1h | Embeding 
#> ⠏ 236/294 ETA:  1h | Embeding 
#> ⠋ 237/294 ETA:  1h | Embeding 
#> ⠙ 238/294 ETA:  1h | Embeding 
#> ⠹ 239/294 ETA:  1h | Embeding 
#> ⠸ 240/294 ETA:  1h | Embeding 
#> ⠼ 241/294 ETA:  1h | Embeding 
#> ⠴ 242/294 ETA:  1h | Embeding 
#> ⠦ 243/294 ETA:  1h | Embeding 
#> ⠧ 244/294 ETA:  1h | Embeding 
#> ⠇ 245/294 ETA:  1h | Embeding 
#> ⠏ 246/294 ETA: 50m | Embeding 
#> ⠋ 247/294 ETA: 49m | Embeding 
#> ⠙ 248/294 ETA: 48m | Embeding 
#> ⠹ 249/294 ETA: 46m | Embeding 
#> ⠸ 250/294 ETA: 45m | Embeding 
#> ⠼ 251/294 ETA: 44m | Embeding 
#> ⠴ 252/294 ETA: 43m | Embeding 
#> ⠦ 253/294 ETA: 42m | Embeding 
#> ⠧ 254/294 ETA: 41m | Embeding 
#> ⠇ 255/294 ETA: 40m | Embeding 
#> ⠏ 256/294 ETA: 39m | Embeding 
#> ⠋ 257/294 ETA: 38m | Embeding 
#> ⠙ 258/294 ETA: 37m | Embeding 
#> ⠹ 259/294 ETA: 36m | Embeding 
#> ⠸ 260/294 ETA: 35m | Embeding 
#> ⠼ 261/294 ETA: 34m | Embeding 
#> ⠴ 262/294 ETA: 33m | Embeding 
#> ⠦ 263/294 ETA: 32m | Embeding 
#> ⠧ 264/294 ETA: 30m | Embeding 
#> ⠇ 265/294 ETA: 29m | Embeding 
#> ⠏ 267/294 ETA: 27m | Embeding 
#> ⠋ 268/294 ETA: 26m | Embeding 
#> ⠙ 269/294 ETA: 25m | Embeding 
#> ⠹ 270/294 ETA: 24m | Embeding 
#> ⠸ 271/294 ETA: 23m | Embeding 
#> ⠼ 272/294 ETA: 22m | Embeding 
#> ⠴ 273/294 ETA: 21m | Embeding 
#> ⠦ 274/294 ETA: 20m | Embeding 
#> ⠧ 275/294 ETA: 19m | Embeding 
#> ⠇ 276/294 ETA: 18m | Embeding 
#> ⠏ 277/294 ETA: 17m | Embeding 
#> ⠋ 278/294 ETA: 16m | Embeding 
#> ⠙ 279/294 ETA: 15m | Embeding 
#> ⠹ 280/294 ETA: 14m | Embeding 
#> ⠸ 282/294 ETA: 12m | Embeding 
#> ⠼ 283/294 ETA: 11m | Embeding 
#> ⠴ 284/294 ETA: 10m | Embeding 
#> ⠦ 285/294 ETA:  9m | Embeding 
#> ⠧ 286/294 ETA:  8m | Embeding 
#> ⠇ 287/294 ETA:  7m | Embeding 
#> ⠏ 288/294 ETA:  6m | Embeding 
#> ⠋ 289/294 ETA:  5m | Embeding 
#> ⠙ 290/294 ETA:  4m | Embeding 
#> ⠹ 291/294 ETA:  3m | Embeding 
#> ⠸ 292/294 ETA:  2m | Embeding 
#> ⠼ 293/294 ETA:  1m | Embeding 
#> ⠼ 294/294 ETA:  0s | Embeding 
#> [1] "/tmp/RtmprQBpaT/OMOP_DATA_FOLDER/oa_my_omop_assistant.duckdb"

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)
#> ℹ Set system prompt.
#> Using model = "gemini-2.5-flash".
#> Error in default_google_credentials(api_key, gemini = TRUE): The package "gargle" is required for Google authentication

chat$chat("How to create an acetaminophen cohort?")
#> Error: object 'chat' not found
# }