Create an ellmer chat with the data of the omop trained model

Create an ellmer chat with the data of the omop trained model

Usage

ellmerChat(name = "omop_assistant", chat = ellmer::chat_google_gemini(), ...)

Arguments

name

Store name.

chat

A chat object.

...

Arguments passed to ragnar::ragnar_register_tool_retrieve()

Value

The chat object with the trained store and prompt.

Examples

# \donttest{
storeCreate(
  embed = \(x) ragnar::embed_ollama(x, model = "mxbai-embed-large"),
  name = "my_omop_assistant"
)
#> ℹ `OMOP_DATA_FOLDER` environment variable is not set, using temp directory.
#> Error in req_perform(req): Failed to perform HTTP request.
#> Caused by error in `curl::curl_fetch_memory()`:
#> ! Couldn't connect to server [localhost]:
#> Failed to connect to localhost port 11434 after 0 ms: Couldn't connect to server

chat <- ellmerChat(
  name = "my_omop_assistant",
  chat = ellmer::chat_google_gemini(),
  top_k = 10L
)
#> ℹ Set system prompt.
#> Using model = "gemini-2.5-flash".
#> Error in default_google_credentials(api_key, gemini = TRUE): The package "gargle" is required for Google authentication

chat$chat("How to create an acetaminophen cohort?")
#> Error: object 'chat' not found
# }